Implemented ios based I/O for DecisionTable
Implemented a rough and ready I/O for rules and decision tables and a
few tests.

Also changed the meaning of operator== and != to mean true comparison of
rules and rule inputs.

Introduced FindFirstMatch function to be used instead of std::find
because we match on RuleInput rather than Rule and operator== is now
reserved for true Rule comparison (as opposed to match lookups).